+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   - In `CorpusListing`:
     - Merged `getContextQueryString` and `getContextQueryStringFromCorpusId` as `getContextParams`
     - Split `getAttributeGroups`/`getStatsAttributeGroups` into `getAttributeGroups`, `getAttributeGroupsCompare`, `getAttributeGroupsExtended` and `getAttributeGroupsStatistics`
+- The JSON button is a regular button, not an image, and downloads the stored, actual response instead of sending a new request
 
 ### Fixed